feat: ✨ Remove harpoon marked files keymap
Previously, harpoon files could be marked but couldn't be removed using
a keybinding. This commit adds that function. Moreover, the harpoon
marks have been extended from only 4 to 8. So a total of 8 files can be
marked and navigated to easily using keybindings.

Another bug that existed with previous harpoon configuration was that it
overrode the key binding for vertical navigation using C-U. This has
been changed. Now harpoon marked fixed are navigated using <C-#> where #
is a number between 1 to 8 inclusive.