feat: ✨ Alacritty configs updated
Alacritty opacity increased from 0.8 to 0.95. This was done to make the
terminal less distracting. Working in the terminal with background
seeping through makes it incredibly hard to focus. Although it looks
visually stunning, it's not practical and functional.

The font size has also increased from 12 to 14. This was done to make it
easier on the eyes to read what's in the terminal.